doc: set MIN_WEST_VERSION to 1.0.0
If west is installed, we have to build the Zephyr documentation with a
recent version of west. This is because the west API documentation is
part of the Zephyr documentation, and the west API documentation
contains Sphinx autodoc directives which pull API documentation out of
the west source code itself. This is similar to how we pull Doxygen
comments out of the Zephyr API headers using breathe. (If west is not
installed, we don't build the west API docs, so you can uninstall west
if you don't want this.)

The current API docs require west v1.0.0 or later to build, since that
version of west introduced new APIs and documentation for them not
available in prior versions.

(cherry picked from commit 523de0d9be8020b305d86c7b3cabf07c5342841d)

Original-Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: 523de0d9be8020b305d86c7b3cabf07c5342841d
Change-Id: Ic81a634b076a8fc6dbef945c29cb9a0206e72770
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4322275
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>