tests: drivers: adc_dma: fix buffer state not being verified
Previously the contents of buffers after an ADC DMA read was
simply printed, but not verified with an zassert that the
value was updated.

This commit updates it to be similar to the adc_api test that
fills the buffer initially with a known value, which is then
used to ensure the ADC DMA functioned successfully.

(cherry picked from commit 7e4de9363f8a6edb1e7b92c0a52c4aa650255f20)

Original-Signed-off-by: Hein Wessels <heinwessels93@gmail.com>
GitOrigin-RevId: 7e4de9363f8a6edb1e7b92c0a52c4aa650255f20
Change-Id: I2b583420a9ed00c4c9823ffa463f8717a5bbc75f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4356815
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>