nrf52_bsim: Fix unfreed memory
There was a chunk of unfreed memory left over at exit().
Clear it as it should be.
The issue is only a cause for a warning in valgrind,
as Linux frees all process booked memory on exit,
but nonetheless it is better fixing it.

(cherry picked from commit bb6fa5707f77d72307008405f19dd83247405392)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: bb6fa5707f77d72307008405f19dd83247405392
Change-Id: If15b22f79b1593d0510a8bc29db570f6be0405b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4336209
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>