drivers: gpio: xlnx: Add interrupt support
Added support for interrupt-driven GPIO for this driver if an interrupt
line is defined. Note that interrupt support is optional in the GPIO
core so it may or may not be present.

The core provides a single state change notification whenever any line
changes state on a given bank, so we need to check the states of the
lines ourselves to determine what has actually changed.

Also, the driver treats the first and second GPIO banks as mostly
separate devices, but a single interrupt is shared by both banks, so the
ISR ends up being registered by the first bank which also triggers
handling on the second bank if required.

(cherry picked from commit ca3a3c1783a8f33460f03373c9c40ee4db2f3387)

Original-Signed-off-by: Robert Hancock <robert.hancock@calian.com>
GitOrigin-RevId: ca3a3c1783a8f33460f03373c9c40ee4db2f3387
Change-Id: If414006e9d383e453aebc8dfa93dd6216702fc6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4572377
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>