z_sched_init: don't use arch_num_cpus()
The reason for arch_num_cpus() is to be able to dynamically adapt to
the actual number of available CPUs at run time.

In the z_sched_init() case, it is not the number of active CPUs that
we need but rather the total number of potential CPUs, and that is
represented by CONFIG_MP_MAX_NUM_CPUS not arch_num_cpus().

(cherry picked from commit 907eea07f23ba1a10544cdad868427b84d863116)

Original-Signed-off-by: Nicolas Pitre <npitre@baylibre.com>
GitOrigin-RevId: 907eea07f23ba1a10544cdad868427b84d863116
Change-Id: Ib34004d1cdc872f5115f24848ac66e4ed9395318
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4397830
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>