tests bsim bap_unicast_audio: Avoid randomness failure
The current seed causes a failure with the updated HW models
due to a change in the random draw, that happens to unluckily
cause the test to fail.
Avoide it with a different random seed.
This is due to a known issue in the controller, which is
being worked on by the maintainer.

(cherry picked from commit ef6c91411c44243e6c1a1647bc100322147889c8)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: ef6c91411c44243e6c1a1647bc100322147889c8
Change-Id: If0c6ff6b1969af5ae61d0098c6fa897d9f03a494
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4360205
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>