rtio: Properly track last sqe in the queue
The pending_sqe logic to track where in the ring queue the concurrent
executor had left off was slightly flawed. It didn't account for starting
all sqes in the queue and ending back up at the beginning.

Instead track the last SQE in the queue, from which the next one in the
queue will the one to start next.

If we happen to sweep the last known SQE in the queue, reset it to NULL
so the next time prepare is called we start at the beginning of the queue
again.

(cherry picked from commit 610d307fa00f4c713aa012f4ae0e6b8bd03c1828)

Original-Signed-off-by: Tom Burdick <thomas.burdick@intel.com>
GitOrigin-RevId: 610d307fa00f4c713aa012f4ae0e6b8bd03c1828
Change-Id: I262a7fc5fc3d3e95e40cc1a1743f15137c7ca574
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4356692
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>