Bluetooth: CSIP: Add support for optional characteristics
Whenever a CSIS instance is registered with
bt_csip_set_member_register we now properly check if the
set size, rank and lock characteristics should be present.
If not, then remove them from the service declaration.

It is done this way, as it is easier (read: possible) to
remove characteristics from the service declarations,
than it is to add the optional characteristics, as the
optional characteristics may be present in some CSIS
instances, but not in others.

(cherry picked from commit 3f8daa676249475141bf1eac38f51c1bcd1d16e5)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 3f8daa676249475141bf1eac38f51c1bcd1d16e5
Change-Id: I53cbd56da7d0418fb99ba63b2e93ad37252dc342
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4521258
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>