Bluetooth: audio: ascs: Cleanup ASCS on callback unregister
This adds cleanup of ASCS that is called when upper layer unregisters
it's callbacks. Without callbacks registered, the service is unusable.
The ASE's that were in non-idle state stay in this state, because the
implementation returns an error on every operation that is performed by
the client. The cleanup added moves all the ASE's to idle state.

(cherry picked from commit 31aa555890ad9377e5dfa0d606bf66d93513681c)

Original-Fixes: #56111
Original-Signed-off-by: Mariusz Skamra <mariusz.skamra@codecoup.pl>
GitOrigin-RevId: 31aa555890ad9377e5dfa0d606bf66d93513681c
Change-Id: I358a86d2cc6e7e99f80ae5165b780708f8dd89ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4378470
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>