bluetooth: host: add checks for connection types
Fail gracefully if an HCI event of one type arrives for a handle of a
different connection type.  The requested types are currently based on
what fields are used, not on the usage context, in order to keep every
correct use so far still working.

A warning is logged if the connection identified by the handle does not
match the requested connection type.

(cherry picked from commit 0885416b602acef926806581626f36803674142f)

Original-Signed-off-by: Arkadiusz Kozdra <akozdra@antmicro.com>
GitOrigin-RevId: 0885416b602acef926806581626f36803674142f
Change-Id: I7ecc25ca6a1e035391dfbc4d716e6d753e4d71d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4667079
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>