Bluetooth: audio: ascs: Fix ASE leak if codec config failed
This fixes ASE that is not returned to pool when Codec Config
operation failed. The ASE is allocated in Codec Config operation
context. With this fix, if the operation failed to complete,
the ASE is returned back to the pool, so that it can be allocated
again by this or other client.

(cherry picked from commit e9a878c0723653015571fe429dd7cba0543f24b7)

Original-Fixes: #56138
Original-Signed-off-by: Mariusz Skamra <mariusz.skamra@codecoup.pl>
GitOrigin-RevId: e9a878c0723653015571fe429dd7cba0543f24b7
Change-Id: Ia8ff387cf13a8e403779c59e8f136330eaaec785
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4378473
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>