west: sign: Add Kconfig with application version
Adds a Kconfig option which controls the version of the application
to use when the image is signed using imgtool.

When an application VERSION file is present, the default value will be
identical to the application version, else it will be 0.0.0+0, but a
project may still decide another value, if it so prefers.

(cherry picked from commit 49389b546a4ab6194d0889f55a2fe00903230810)

Original-Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>
Original-Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
GitOrigin-RevId: 49389b546a4ab6194d0889f55a2fe00903230810
Change-Id: Ia266f16831dbc6d8f7ac622bc0cf1f27c4b06416
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4545566
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>