FROMPULL: cmake: python: Allow for explicit specification of a Python interpreter
Setting PYTHON_PREFER may not select the desired interpreter under all
circumstances.  find_package(Python3 ... EXACT) will try to find the Python
interpreter of the desired version, but will not provide the correct
interpreter if it's not in PATH, or if multiple interpreters exist of the
same major+minor version in the PATH.

Let's add an option for a user to explicitly specify PYTHON_EXECUTABLE, not
just a preference.  The build system will then try that executable, and
only that executable, and if it doesn't meet the version requirement, it'll
error out instead of keep searching.

FROMPULL request:
https://github.com/zephyrproject-rtos/zephyr/pull/56205

BUG=b:269281716
TEST=build with bazel

Change-Id: I26dee721ebc16bd4e9cbd49cb48c170d7a8ab1eb
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4615502
Reviewed-by: Aaron Massey <aaronmassey@google.com>