pm: device_runtime: PM unsupported is not an error
Returning an error code when PM is not supported for a device only makes
writing drivers harder, as instead of checking for failures with `< 0`,
they also need to check for `-ENOTSUP`.

From the point of view of a driver, an `-ENOTSUP` return value is
equivalent to success, as if the device it is trying to turn on doesn't
support PM, by definition it is already enabled. This is equivalent to
the API behaviour when `CONFIG_PM_DEVICE_RUNTIME=n`.

Whether a device supports PM or not can still be determined at runtime
by inspecting the return code of `pm_device_runtime_enable` if
necessary.

(cherry picked from commit c6191858c368bca9325a70a34ca4977cf4aa407c)

Original-Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
GitOrigin-RevId: c6191858c368bca9325a70a34ca4977cf4aa407c
Change-Id: I790cfd867775932dff8f260927c127f118a94663
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4381135
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>