Bluetooth: BAP: Store biginfo data in broadcast_sink->codec_qos
Store information from the biginfo in a
broadcast_sink->codec_qos struct, and add a reference to that struct
when syncing, so that the stream->qos contains a valid struct
with properly initialized information.

This also adds a guard for the codec_qos.latency as that
is not set nor used by the broadcast sink (as it is
not available).

(cherry picked from commit 630f6a7d82a117e94f88604ab7593394d4e4fda6)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 630f6a7d82a117e94f88604ab7593394d4e4fda6
Change-Id: I61919183ea5e51eedc8030e6fe6951af46440aec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4384982
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>