test: misc: add a test application for check_init_priorities
Add a test specifically for the check_init_priorities scripts.

The test creates a set of fake devices to exercise the three possible
conditions, run a build, run the check script and validate the script
output.

The check is meant to fail the build on error but that's bypassed in
this case as that would fail the twister run. That specific bit is
covered in unit tests anyway.

(cherry picked from commit 2354cc9e255038768ffddd655e38e1425d16324d)

Original-Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
GitOrigin-RevId: 2354cc9e255038768ffddd655e38e1425d16324d
Change-Id: I743388622a6e602f4e9e819db7f505bb88e5c3b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4570748
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>