samples: zbus: fix variable sizes and simplify benchmark sample
The benchmark sample was not suitable for constrained devices.
With u32, the maximum accumulated time was about 4 seconds.
Some devices take more than 4 seconds to finish the benchmark.
This commit fixes the time counting by replacing the variables
and calls to 64 bits and removes all the dynamic allocation
memory on the sample to keep that as simple as possible. It
adjusts the documentation and the tests as well.

(cherry picked from commit a24c307baafaab344997cf33ce7be9ba9cea06ae)

Original-Signed-off-by: Rodrigo Peixoto <rodrigopex@gmail.com>
GitOrigin-RevId: a24c307baafaab344997cf33ce7be9ba9cea06ae
Change-Id: Ice7ce74d44c963c0470c670b7accede9df0dc576
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4430233
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>