Bluetooth: Tests: Check CCC Store
The first test check that if a bonded peer subscribe to a CCC, this
change is correctly stored to flash.

The second test, which is currently failing, is included to reproduce a
bug that occur when CCC or CF store on write is disabled and the other
one is enabled, leading to the delayed store being enabled. Causing the
value of the corresponding `n` selected option to not be stored at all.

(cherry picked from commit f4ecc4e102a2b9433ce37450578c1d9dbeea3f18)

Original-Signed-off-by: Théo Battrel <theo.battrel@nordicsemi.no>
GitOrigin-RevId: f4ecc4e102a2b9433ce37450578c1d9dbeea3f18
Change-Id: I3466228b110f5eb2cbee9c234bf3a3826828de2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4585176
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>