dts/bindings/rtc/rtc-device.yaml: alarms-count optional
This commit makes the alarms-count dts property of the
rtc-device.yaml optional, setting the default to 0.

This simplifies the dts rtc dts node by not requiring
the property to be set to 0 if it is not supported.

(cherry picked from commit 9eae6fda9930eb1ae674d02f6992dbbc274fbd2a)

Original-Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>
GitOrigin-RevId: 9eae6fda9930eb1ae674d02f6992dbbc274fbd2a
Change-Id: I0dc73b166c2e466ed90bbd88ac54c37fb5dfd4eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4454126
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>