boards: arm64: mimx93_evk: Introduce new board for SOF usage
This commit introduces a new variant of the mimx93_evk_a55 board
used to run the SOF module.

This is required because we don't want to pollute the imx93_a55
generic board with SOF-specific nodes such as the memory nodes.
Also, the SOF-specific drivers should not be activated in the
generic board.

SOF with Jailhouse also requires some specific configurations
such as the 40-bit PAs and VAs and the static data cache line
size so this is another reason for which this new
board had to be created.

(cherry picked from commit 1b923275e64ae7b6650c71a1f3e7ef6180f0e273)

Original-Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
GitOrigin-RevId: 1b923275e64ae7b6650c71a1f3e7ef6180f0e273
Change-Id: I9af65df8aa386e2cabcb81017797fef52e3f5a77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4446744
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>