tests: drivers: build_all: add build for wdt_xilinx_axi
There is currently not an in-tree Zephyr board which enables the
wdt_xilinx_axi driver as part of its standard configuration. (It can be
enabled fairly easily with the Digilent Arty board, but that requires
modifying the ARM-provided FPGA block design in Vivado and rebuilding
the FPGA image, so it cannot be used with the default FPGA image.)

Add a build_all target to provide build coverage for this driver.

(cherry picked from commit 9988067c5aeef5f14facd5f079fe64facf578b2c)

Original-Signed-off-by: Robert Hancock <robert.hancock@calian.com>
GitOrigin-RevId: 9988067c5aeef5f14facd5f079fe64facf578b2c
Change-Id: I48802f7cafb9e7a168ab862d16e8f8c14ae99633
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4550567
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>