drivers: watchdog: add Xilinx AXI Timebase WDT driver
Add a driver for the Xilinx AXI Timebase WDT logic core. This can be
instantiated on various Xilinx FPGA-based platforms such as the
Digilent Arty, although it is not part of the default image used with
the Zephyr board configuration.

The driver can also optionally implement the HWINFO API to allow
determining whether the last system reset was initiated by the WDT.
Since this is a standalone IP core which could be used a variety of
configurations, this support is optional in case the system/SoC it is
used with already implements this support.

(cherry picked from commit e57bab59ff732e0134f5da94ded407b883fc41f8)

Original-Signed-off-by: Robert Hancock <robert.hancock@calian.com>
GitOrigin-RevId: e57bab59ff732e0134f5da94ded407b883fc41f8
Change-Id: I4f935981d85483937d47534146b7eb7a9e740011
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4550566
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>