rtio: rework RTIO mempool management
The previous RTIO mempool allocation was found to be faulty. Since the
map entries were indexed using the sqe, as soon as the sqe and cqe were
no longer in sync, there was a contention. Effectively a new sqe needing
memory would assume the entry was already consumed by themselves. This
was caused by 'rtio_sqe_rx_buf()' being effectively reentrant (i.e. if
called more than once, it'll return the same buffer). So when the map
entry was marked as allocated it was assumed that it belonged to the sqe
in question and used. This led to the producer writing to memory that
was actively being read by the consumer.

Instead, the sqe's own buf/buf_len fields are used to store the memory.
Once allocated, the memory is stored there and no mapping is needed.

(cherry picked from commit 00a7251613c11d1f85269f3bbd7a199cfa9071d5)

Original-Signed-off-by: Yuval Peress <peress@google.com>
GitOrigin-RevId: 00a7251613c11d1f85269f3bbd7a199cfa9071d5
Change-Id: I50d505893b557a35c02b1e79e981384b55bf1728
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4498597
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>