logging: Fixed BLE backend buffer size independent of MTU
The BLE backend has a bug that when the MTU size is less than
the buffer the notification always fails and does not allow
to flush the buffer.

This fix checks for the MTU size of the current connection
and adjusts it. In addition, the buffer size is no longer
settable by the user and depends on the transmission size
of MTU set with CONFIG_BT_L2CAP_TX_MTU.

(cherry picked from commit bee43bd972d2174ca624f08207bc8ed92c8f9edb)

Original-Signed-off-by: Victor Chavez <chavez-bermudez@fh-aachen.de>
GitOrigin-RevId: bee43bd972d2174ca624f08207bc8ed92c8f9edb
Change-Id: I81d6bd29429e39e5585423eb7a3d7c1237b09ec7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4626677
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>