boards: particle_boron: Fix init of modem pins
The Boron has a power switch that controls power to a level shifter
for the modem's serial line. Previously, this switch was being enabled
after the modem driver had already started. It is now controlled using
a fixed regulator.

The board initialization was also configuring the V_INT_DETECT pin.
However, this pin is already handled as mdm-vint-gpios in the dts when
the ublox-sara-r4 driver is used. Since it is not used otherwise,
there is no need to configure it in the board init.

(cherry picked from commit b6c80d16a1ce9195dfeae62b237df9a2b270dd54)

Original-Fixes: #56883
Original-Signed-off-by: Göran Weinholt <goran@weinholt.se>
GitOrigin-RevId: b6c80d16a1ce9195dfeae62b237df9a2b270dd54
Change-Id: Icc027eb32f785111c88a1dbd935323b015ae39a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4576991
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>