nrf52_bsim: Do not load Tsymbols from main
We do not use them for anything yet, but they do
use a lot of heap, specially for the networking apps
with openthread.
If we need them later we can always add it again.

Moreover, there is an issue is this bsim component
detected by ASAN when loading symbols with too long names,
the issue is fixed in the master branch, but better
if users do not need to update for a component which
is not used.

(cherry picked from commit 3a8589eadb9d355860eb823879857178816c6982)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 3a8589eadb9d355860eb823879857178816c6982
Change-Id: Idc60b0daa4f09f4791d5bd867456ec5c7810c324
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4462662
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>